clean up file structure and rename error map
